create partition return PartitionResponse and drop partition return boolean
